requirements.txt
numpy<2
pandas
matplotlib
scikit-learn
faiss-cpu==1.7.4